fix bug in atof
When called with arguments with long precision, atof
sometomes was return incorrect results. This was due
integer overflow in small_atoi() subcalls.
Now calls to small_atoi() replaced with calls to
small_atoll(), which operates with long long ints.
This eliminates problem with atof in all allowed
precision range (up to 10).